+++ pybuoy/api/forecasts.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 from datetime import datetime as dt
 from json import loads
 from os.path import dirname, join
@@ -7,24 +8,26 @@ import lxml.etree as ET
 
 from pybuoy.api.base import ApiBase
 from pybuoy.const import API_PATH, Endpoints
-
-# from pybuoy.observation.observation import MeteorologicalPrediction
-# from pybuoy.observation.observations import MeteorologicalPredictions
-# from pybuoy.unit_mappings import MeteorologicalKey
+from pybuoy.observation import ForecastObservation, ForecastObservations
 
 
 class Forecasts(ApiBase):
     # https://graphical.weather.gov/xml/rest.php
-    def get(self, lat: float, lon: float, beginDate: str, endDate: str):
+    def get(self, lat: float, lon: float, start_date: str, end_date: str):
         dwml_data = self.forecast_by_lat_lon(
             lat=lat,
             lon=lon,
-            start_date=dt.fromisoformat(beginDate),
-            end_date=dt.fromisoformat(endDate),
+            start_date=dt.fromisoformat(start_date),
+            end_date=dt.fromisoformat(end_date),
         )
+        grouped = self.__group_parameters_by_time(dwml_data["data"]["parameters"])
         # TODO: serialize into python objects for end-users
-        # MeteorologicalPredictions(observations=dwml_data)
-        return dwml_data
+        predictions = [
+            ForecastObservation(grouped[key], dt.fromisoformat(key))
+            for key in grouped.keys()
+        ]
+        parsed = ForecastObservations(observations=predictions, repr_limit=10)
+        return parsed
 
     def forecast_by_lat_lon(
         self,
@@ -86,3 +89,45 @@ class Forecasts(ApiBase):
 
         json_load = loads(str(x))
         return json_load
+
+    # TODO: should XSLT flatten data instead
+    # TODO: add documentation, if this will be merged
+    def __group_parameters_by_time(self, parameters: dict):
+        # TODO: refactor mapping
+        ele_map = {
+            "direction": ("wind-direction",),
+            "wind": (
+                "gust",
+                "sustained",
+            ),
+            "water-state": ("significant",),
+        }
+        default_val = {"value": "nan"}
+        # TODO: initialize defaults dynamically on _init_
+        groupedby_timestamp: dict[str, dict] = defaultdict(
+            lambda: defaultdict(
+                dict,
+                wave_height=default_val,
+                wind_direction=default_val,
+                wind_speed_gust=default_val,
+                wind_speed=default_val,
+            )
+        )
+        for k, v in parameters.items():
+            for key in v.keys():
+                if key in ele_map[k]:
+                    for period in parameters[k][key]["periods"]:
+                        timestamp = period["start-time"]
+                        new_key = "_".join(
+                            parameters[k][key]["name"].lower().split(" ")
+                        )
+                        # TODO: eval ForecastObservation
+                        groupedby_timestamp[timestamp][new_key] = {
+                            "value": period["value"],
+                            # TODO: transform to proper label here
+                            "label": parameters[k][key]["name"],
+                            "_units": parameters[k][key]["units"],
+                            "_k": k,
+                        }
+        # TODO: document structure
+        return groupedby_timestamp