+++ pybuoy/unit_mappings.py
@@ -11,6 +11,13 @@ class BaseKey(Enum):
     ...
 
 
+class ForecastKey(BaseKey):
+    wind_direction = "Wind Direction"
+    wave_height = "Wave Height"
+    wind_speed = "Wind Speed"
+    wind_speed_gust = "Wind Speed Gust"
+
+
 class MeteorologicalKey(BaseKey):
     WDIR = "WDIR"
     WSPD = "WSPD"
@@ -106,11 +113,11 @@ WAVE_SUMMARY: dict[WaveSummaryKey, MeasurementsAndUnits] = {
     },  # direction of waves at dominant period (DPD)
 }
 
-FORECAST: dict[MeteorologicalKey, MeasurementsAndUnits] = {
-    MeteorologicalKey.WDIR: METEOROLOGICAL[MeteorologicalKey.WDIR],
-    MeteorologicalKey.WSPD: {"label": "Wind Speed", "unit": "knots"},
-    MeteorologicalKey.GST: {"label": "Wind Gust", "unit": "knots"},
-    MeteorologicalKey.WVHT: {"label": "Wave Height", "unit": "feet"},
+FORECAST: dict[ForecastKey, MeasurementsAndUnits] = {
+    ForecastKey.wave_height: {"label": "Wave Height", "unit": "feet"},
+    ForecastKey.wind_direction: {"label": "Wind Direction", "unit": "degrees"},
+    ForecastKey.wind_speed: {"label": "Wind Speed", "unit": "knots"},
+    ForecastKey.wind_speed_gust: {"label": "Wind Gust", "unit": "knots"},
 }
 
 # TODO: update below for dynamic mapping